Created a player search component and swapped out the old form on the spellslots page.
